requirements.txt
feedparser==6.0.8
django==3.2.3
djangorestframework==3.13.1
psycopg2-binary==2.8.6
black==22.1.0
djangorestframework-simplejwt==5.0.0
dj-database-url==0.5.0
celery==5.2.3
django-redis==5.2.0
celery-redbeat==2.0.0
pytest-django==4.5.2
factory_boy==3.2.1
pytest-factoryboy==2.1.0
pytest-env==0.6.2
pytest-celery==0.0.0
drf-yasg==1.20.0
